SwiftUI - User Events
It has been modified so that the user can select favorites in the List Details View, adjustments were applied so that the toggle to show favorites was only on the main screen and not in the List Details View. A button was implemented to select a programmer as a favorite and show it on the main screen.